Refactor WebSubUtils
